requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0 